requirements.txt
et-xmlfile==1.1.0
openpyxl==3.0.7
pyserial==3.5
